+++ Gruntfile.js
@@ -1,12 +1,12 @@
 module.exports = function(grunt) {
 	var templateThings = function(){
-			var parent =  grunt.file.read('./src/object.js');
-			var child = grunt.file.read('./src/worker.object.js');
+			var parent =  grunt.file.read('./src/core.js');
+			var child = grunt.file.read('./src/worker.js');
 			var replacedChild = "['"+child.replace(/\$\$(.+?)\$\$/,function(a,b){
 				return "',"+b+",'";
 			}).replace(/\n/gm,'')+"']";
 			var out = parent.replace(/\$\$fObj\$\$/,replacedChild);
-			grunt.file.write('./src/temp.object.js',out);
+			grunt.file.write('./src/temp.js',out);
 	};
 	// Project configuration.
 	grunt.initConfig({
@@ -16,7 +16,7 @@ module.exports = function(grunt) {
 				options:{
 					banner:'/*! <%= pkg.name %> <%= pkg.version %> <%= grunt.template.today("yyyy-mm-dd") %>*/\n/*!(c)2013 Calvin Metcalf @license MIT https://github.com/calvinmetcalf/communist */\n/*!Includes Promiscuous (c)2013 Ruben Verborgh @license MIT https://github.com/RubenVerborgh/promiscuous*/\n/*!Includes Material from setImmediate Copyright (c) 2012 Barnesandnoble.com, llc, Donavon West, and Domenic Denicola @license MIT https://github.com/NobleJS/setImmediate */\n',
 					mangle: {
-						except: ['Communist']
+						except: ['Communist','CommunistQueue','FakeCommunist','Promise','Deferred']
 					}
 				},
 				src: 'dist/<%= pkg.name %>.js',
@@ -31,7 +31,7 @@ module.exports = function(grunt) {
 					seperator:";\n",
 					footer : 'communist.version = "<%= pkg.version %>";\n})(this);}'
 				},
-				files: {'dist/<%= pkg.name %>.js':['src/IE.js','src/setImmediate.js','src/promiscuous.js','src/utils.js','src/fakeWorkers.js','src/temp.object.js','src/queue.js','src/wrapup.js']}
+				files: {'dist/<%= pkg.name %>.js':['src/IE.js','src/setImmediate.js','src/promiscuous.js','src/utils.js','src/fakeWorker.js','src/temp.js','src/queue.js','src/wrapup.js']}
 			}
 		},
 		mocha_phantomjs: {